Pause Menu
Pausing should be (mostly) implemented. It uses a system that the game
controller has a ispaused value and most update functions in the game
check this value before doing anything. This allows the pausing to
happen without messing with the time scale. All game options are
currently accessible in the pause menu and you can see the game graphics
change as you can values (I think it's cool).