Physics Changes
Several physics fixes. Springs now set the player's velocity to their
power instead of applying a force. Wall sticking should now be prevented
by disabling input for a very short time if the player collides with the
environment when they are not on the ground. Made enemies wait if they
are a certain distance away from the player.